Refactor the import of optimal_interpolation
